Update Rake report tasks namespaces
This commit introduces a new `Reports` module to incorporate
all logic pertaining to Rake tasks that produce reports
