+++ spec/lib/reports/content_change_email_status_count_spec.rb
@@ -1,4 +1,4 @@
-RSpec.describe ContentChangeEmailStatusCount do
+RSpec.describe Reports::ContentChangeEmailStatusCount do
   before do
     3.times { create(:email, status: 'sent') }
     2.times { create(:email, status: 'pending') }
